move past consensus code to legacy package
