+++ consensus/legacy/api/parliament_api_test.go
@@ -3,8 +3,8 @@ package api
 import (
 	"testing"
 
-	"github.com/it-chain/it-chain-Engine/consensus/domain/model/parliament"
-	"github.com/it-chain/it-chain-Engine/consensus/domain/repository"
+	"github.com/it-chain/it-chain-Engine/consensus/legacy/domain/model/parliament"
+	"github.com/it-chain/it-chain-Engine/consensus/legacy/domain/repository"
 	"github.com/stretchr/testify/assert"
 )
