+++ consensus/legacy/domain/model/parliament/parliament_test.go
@@ -3,7 +3,7 @@ package parliament
 import (
 	"testing"
 
-	"github.com/it-chain/it-chain-Engine/consensus/domain/model/consensus"
+	"github.com/it-chain/it-chain-Engine/consensus/legacy/domain/model/consensus"
 	"github.com/stretchr/testify/assert"
 )
