+++ consensus/legacy/domain/repository/consensus_repository.go
@@ -1,6 +1,6 @@
 package repository
 
-import "github.com/it-chain/it-chain-Engine/consensus/domain/model/consensus"
+import "github.com/it-chain/it-chain-Engine/consensus/legacy/domain/model/consensus"
 
 type ConsensusRepository interface {
 	Save(consensus consensus.Consensus)
