+++ consensus/legacy/domain/repository/parliament_repository.go
@@ -3,7 +3,7 @@ package repository
 import (
 	"sync"
 
-	"github.com/it-chain/it-chain-Engine/consensus/domain/model/parliament"
+	"github.com/it-chain/it-chain-Engine/consensus/legacy/domain/model/parliament"
 )
 
 type ParlimentRepository interface {
